backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
drf-spectacular==0.28.0
boto3==1.36.2
google-generativeai==0.8.4
